# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "uSocial API",
        "description": "API documentation for uSocial",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/influencers/become": {
            "post": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Become an influencer",
                "operationId": "4b2a956f3fe42f475631c5635f3097a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bio",
                                    "gem_cost_per_dm"
                                ],
                                "properties": {
                                    "bio": {
                                        "description": "The bio of the user",
                                        "type": "string",
                                        "maxLength": 1000
                                    },
                                    "gem_cost_per_dm": {
                                        "description": "Cost of gems per direct message",
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully became an influencer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are now an influencer"
                                        },
                                        "influencer": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Passionate influencer about tech and gadgets"
                                                },
                                                "gem_cost_per_dm": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User is already an influencer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User is already an influencer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "bio": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The bio field is required."
                                                    }
                                                },
                                                "gem_cost_per_dm": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The gem_cost_per_dm must be at least 1."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/influencers": {
            "get": {
                "tags": [
                    "Influencers"
                ],
                "summary": "List all influencers",
                "description": "Retrieve a paginated list of all influencers, excluding the authenticated user's influencer profile if they are an influencer.",
                "operationId": "listInfluencers",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of influencers per page (default is 10).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a paginated list of influencers.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "Current page number.",
                                            "type": "integer"
                                        },
                                        "data": {
                                            "description": "List of influencers.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Influencer ID.",
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "description": "User ID of the influencer.",
                                                        "type": "integer"
                                                    },
                                                    "bio": {
                                                        "description": "Bio of the influencer.",
                                                        "type": "string"
                                                    },
                                                    "gem_cost_per_dm": {
                                                        "description": "Gem cost per direct message.",
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "description": "Timestamp of when the influencer was created.",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "description": "Timestamp of when the influencer was last updated.",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "description": "Total number of influencers.",
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "description": "Number of influencers per page.",
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "description": "Last page number.",
                                            "type": "integer"
                                        },
                                        "next_page_url": {
                                            "description": "URL of the next page.",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "prev_page_url": {
                                            "description": "URL of the previous page.",
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, missing or invalid token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}